# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TM Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /assets:
    get:
      operationId: listAssets
      tags: [assets]
      responses:
        "200":
          description: List of assets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Asset"
    post:
      operationId: createAsset
      tags: [assets]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
  /assets/{id}:
    patch:
      operationId: updateAsset
      tags: [assets]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
    delete:
      operationId: deleteAsset
      tags: [assets]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # Gear Items
  /gear:
    get:
      operationId: listGearItems
      tags: [gear]
      responses:
        "200":
          description: List of gear items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GearItem"
    post:
      operationId: createGearItem
      tags: [gear]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GearItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GearItem"
  /gear/{id}:
    patch:
      operationId: updateGearItem
      tags: [gear]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GearItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GearItem"
    delete:
      operationId: deleteGearItem
      tags: [gear]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # Productions
...
  /build-projects:
    get:
      operationId: listBuildProjects
      tags: [build-projects]
      responses:
        "200":
          description: List of build projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BuildProject"
    post:
      operationId: createBuildProject
      tags: [build-projects]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BuildProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BuildProject"
  /build-projects/{id}:
    patch:
      operationId: updateBuildProject
      tags: [build-projects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BuildProjectUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BuildProject"
    delete:
      operationId: deleteBuildProject
      tags: [build-projects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # Projects
...
  /projects/{projectId}/call-sheets:
    get:
      operationId: listCallSheets
      tags: [call-sheets]
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of call sheets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CallSheet"
    post:
      operationId: createCallSheet
      tags: [call-sheets]
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CallSheetInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallSheet"
  /call-sheets/{id}:
    delete:
      operationId: deleteCallSheet
      tags: [call-sheets]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /call-sheets/{callSheetId}/entries:
    get:
      operationId: listCallSheetEntries
      tags: [call-sheets]
      parameters:
        - name: callSheetId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CallSheetEntry"
    post:
      operationId: createCallSheetEntry
      tags: [call-sheets]
      parameters:
        - name: callSheetId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CallSheetEntryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallSheetEntry"
  /call-sheet-entries/{id}:
    delete:
      operationId: deleteCallSheetEntry
      tags: [call-sheets]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # Script Breakdowns
...
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}:
    patch:
      operationId: updateLead
      tags: [leads]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    delete:
      operationId: deleteLead
      tags: [leads]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # Staging Environments
...
components:
  schemas:
...
    Asset:
      type: object
      required: [id, name, type, status, comments, views, createdAt, updatedAt]
      properties:
        id: { type: string }
        name: { type: string }
        type: { type: string }
        duration: { type: ["string", "null"] }
        size: { type: ["string", "null"] }
        status: { type: string }
        comments: { type: integer }
        views: { type: integer }
        client: { type: ["string", "null"] }
        userId: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    AssetInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        type: { type: string }
        duration: { type: string }
        size: { type: string }
        status: { type: string }
        client: { type: string }
    AssetUpdate:
      type: object
      properties:
        name: { type: string }
        type: { type: string }
        duration: { type: string }
        size: { type: string }
        status: { type: string }
        client: { type: string }
    GearItem:
      type: object
      required: [id, name, category, status, condition, createdAt, updatedAt]
      properties:
        id: { type: string }
        name: { type: string }
        category: { type: string }
        status: { type: string }
        location: { type: ["string", "null"] }
        lastUsed: { type: ["string", "null"] }
        condition: { type: string }
        reservedFor: { type: ["string", "null"] }
        userId: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    GearItemInput:
      type: object
      required: [name, category]
      properties:
        name: { type: string }
        category: { type: string }
        status: { type: string }
        location: { type: string }
        lastUsed: { type: string }
        condition: { type: string }
        reservedFor: { type: string }
    GearItemUpdate:
      type: object
      properties:
        name: { type: string }
        category: { type: string }
        status: { type: string }
        location: { type: string }
        lastUsed: { type: string }
        condition: { type: string }
        reservedFor: { type: string }
...
    BuildProject:
      type: object
      required: [id, name, status, feedback, branch, progress, createdAt, updatedAt]
      properties:
        id: { type: string }
        name: { type: string }
        url: { type: ["string", "null"] }
        status: { type: string }
        lastDeploy: { type: ["string", "null"] }
        feedback: { type: integer }
        branch: { type: string }
        progress: { type: integer }
        userId: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    BuildProjectInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        url: { type: string }
        status: { type: string }
        branch: { type: string }
        progress: { type: integer }
    BuildProjectUpdate:
      type: object
      properties:
        name: { type: string }
        url: { type: string }
        status: { type: string }
        lastDeploy: { type: string }
        feedback: { type: integer }
        branch: { type: string }
        progress: { type: integer }
...
    CallSheet:
      type: object
      required: [id, projectId, userId, shootDate, callTime, createdAt, updatedAt]
      properties:
        id: { type: string }
        projectId: { type: string }
        userId: { type: string }
        shootDate: { type: string }
        callTime: { type: string }
        location: { type: ["string", "null"] }
        weatherNotes: { type: ["string", "null"] }
        generalNotes: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    CallSheetInput:
      type: object
      required: [shootDate]
      properties:
        shootDate: { type: string }
        callTime: { type: string }
        location: { type: string }
        weatherNotes: { type: string }
        generalNotes: { type: string }
    CallSheetEntry:
      type: object
      required: [id, callSheetId, userId, personName, role, callTime, createdAt]
      properties:
        id: { type: string }
        callSheetId: { type: string }
        userId: { type: string }
        personName: { type: string }
        role: { type: string }
        callTime: { type: string }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
    CallSheetEntryInput:
      type: object
      required: [personName, role]
      properties:
        personName: { type: string }
        role: { type: string }
        callTime: { type: string }
        notes: { type: string }
...
    Lead:
      type: object
      required: [id, userId, name, company, value, currency, service, stage, createdAt, updatedAt]
      properties:
        id: { type: string }
        userId: { type: string }
        name: { type: string }
        company: { type: string }
        value: { type: string }
        currency: { type: string }
        service: { type: string }
        stage: { type: string }
        notes: { type: ["string", "null"] }
        projectId: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    LeadInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        company: { type: string }
        value: { type: string }
        currency: { type: string }
        service: { type: string }
        stage: { type: string }
        notes: { type: string }
        projectId: { type: string }
    LeadUpdate:
      type: object
      properties:
        name: { type: string }
        company: { type: string }
        value: { type: string }
        currency: { type: string }
        service: { type: string }
        stage: { type: string }
        notes: { type: string }
        projectId: { type: string }
...