# lib/api-zod/src/generated/types/asset.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TM Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface Asset {
  id: string;
  name: string;
  type: string;
  /** @nullable */
  duration?: string | null;
  /** @nullable */
  size?: string | null;
  status: string;
  comments: number;
  views: number;
  /** @nullable */
  client?: string | null;
  /** @nullable */
  userId?: string | null;
  createdAt: string;
  updatedAt: string;
}